Clean up Sigma controller sign comparisons
Validate signed ZIMH callback values before comparing them with unsigned
controller type state. Add a direct unit test for rejecting a negative
disk controller type before it can be stored in unsigned state.

Also compare the terminal panel interrupt character after the input
character has been masked into the same domain.

This removes the related -Wsign-compare warnings.

+++ simulators/sigma/sigma_tt.c
@@ -191,7 +191,7 @@ if (c & SCPE_BREAK) {                                   /* break? */
     return SCPE_OK;
     }
 c = c & 0x7F;
-if (c == tti_panel)                                     /* panel interrupt? */
+if (c == (int32) tti_panel)                             /* panel interrupt? */
     return io_set_pint ();
 uptr->pos = uptr->pos + 1;                              /* incr count */
 if (c == '\r')                                          /* map CR to NL */
